fix: use correct: target offset
We differentiate between LFS shared and executable, but
this is done earlier in the code path. What happens is that
when target is shared obejct the e_entry points to
lsb_so_addr variable, which is in fact the actual
target offset, this is why the destination
address was OK to have lsb_so_addr.

However it was confunsing to have that when we write and with this
commit it is fixed by using the more clear target_offset variable.